requirements.txt
python-consul @ git+https://github.com/cablehead/python-consul.git
grpcio
grpcio-tools